Fix prepareNotification data merging
Currently prepareNotification use the array concatenation operator. By using this operator, we can't override the initial notification data with the params. So for example, the PSR4 adapter does not work well, when we want to send an error it's an info.
We can fix it by using the array_merge function.